requirements.txt
discord.py
aiosqlite
urlextract
python-twitter
